- separated all classes and interfaces into files - Use AsyncIteratableIterator class instead of IteratableIterator - renamed IScraperConfiguration to IScraper
