+++ package.json
@@ -14,31 +14,31 @@
   },
   "scripts": {
     "build": "tsc",
-    "link": "cd dist && npm link",
+    "link": "cd dist && yarn link",
     "dev": "nodemon .\\src\\Example.ts",
     "lint": "tslint --project tsconfig.json && yarn typecheck",
     "typecheck": "tsc --noEmit",
     "test": "mocha -r ts-node/register test/**/*.test.ts",
-    "prepare": "npm run build"
+    "prepare": "yarn run build"
   },
   "bugs": {
     "url": "https://github.com/demokratie-live/scapacra/issues"
   },
   "devDependencies": {
-    "@types/chai": "^4.1.7",
-    "@types/mocha": "^5.2.5",
-    "@types/node": "^10.12.18",
+    "@types/chai": "^4.2.3",
+    "@types/mocha": "^5.2.7",
+    "@types/node": "^12.7.7",
     "chai": "^4.2.0",
-    "mocha": "^5.2.0",
-    "nodemon": "^1.18.6",
-    "ts-node": "^7.0.1",
-    "tslint": "^5.12.0"
+    "mocha": "^6.2.0",
+    "nodemon": "^1.19.2",
+    "ts-node": "^8.4.1",
+    "tslint": "^5.20.0"
   },
   "dependencies": {
-    "@types/xml2js": "^0.4.3",
+    "@types/xml2js": "^0.4.5",
     "@types/xmldom": "^0.1.29",
-    "typescript": "^3.1.6",
-    "xml2js": "^0.4.19",
+    "typescript": "^3.6.3",
+    "xml2js": "^0.4.22",
     "xmldom": "^0.1.27",
     "xpath": "0.0.27"
   }
