+++ src/index.ts
@@ -1,2 +1,8 @@
-export { IDataPackage, IMetadata, DataType, IBrowser, IParser, IScraperConfiguration, Scraper } from './Importer';
+export { IDataPackage } from './IDataPackage';
+export { IMetadata } from './IMetadata';
+export { DataType } from './DataType';
+export { IBrowser } from './IBrowser';
+export { IParser } from './IParser';
+export { IScraper } from './IScraper';
+export { Scraper } from './Scraper';
 export { DocumentEvaluater } from './evaluator/DocumentEvaluator';
